ListAndDictionary homework

+++ Models/Models/Employee.cs
@@ -5,14 +5,8 @@
 /// </summary>
 public class Employee : Person
 {
-    /// <summary>
-    /// Контракт.
-    /// </summary>
     public string Contract { get; set; }
     
-    /// <summary>
-    /// Зарплата.
-    /// </summary>
-    public int Salary { get; set; }
+    public decimal Salary { get; set; }
     
 }
